Finish exercise 5.11: Add windows confirm to delete blog

+++ part5/blog-list/bloglist-frontend/src/components/Blog.jsx
@@ -25,7 +25,9 @@ const Blog = ({ blog, handleLike, handleDelete }) => {
 
   const removeBlog = () => {
     console.log('removing blog...')
-    handleDelete(blog.id)
+    if (window.confirm(`Delete ${blog.title}?`)) {
+      handleDelete(blog.id)
+    }
   }
 
   return (
